fix: ActionChange dto에 값 삽입시 Long 타입 맞춰주기

+++ src/main/java/com/example/wini/domain/note/repository/NoteCustomRepositoryImpl.java
@@ -78,7 +78,7 @@ public class NoteCustomRepositoryImpl implements NoteCustomRepository {
         NumberExpression<Long> increaseCount = thisMonthNotes.subtract(lastMonthNotes);
 
         return queryFactory
-                .select(new QActionChange(action, increaseCount))
+                .select(new QActionChange(action, increaseCount.longValue()))
                 .from(note)
                 .join(note.action, action)
                 .join(action.actionCategory, actionCategory)
@@ -98,7 +98,7 @@ public class NoteCustomRepositoryImpl implements NoteCustomRepository {
         NumberExpression<Long> decreaseCount = thisMonthNotes.subtract(lastMonthNotes);
 
         return queryFactory
-                .select(new QActionChange(action, decreaseCount))
+                .select(new QActionChange(action, decreaseCount.longValue()))
                 .from(note)
                 .join(note.action, action)
                 .join(action.actionCategory, actionCategory)
